This is for UD-476 Eliminate UUID of network from config fiile for STRING loader
The STRING Loader is modified to get network summaries for the current user from the server
and check if network with the network that is generated already exists.  If yes, than
the existing network on the server is updated.  If no, then a new network is created on the server.